+++ ubuntu_clipboard/ui/__init__.py
@@ -15,6 +15,7 @@ try:  # pragma: no cover - depends on the environment
     import gi
 
     gi.require_version("Gtk", "4.0")
+    gi.require_version("Gdk", "4.0")
     gi.require_version("Pango", "1.0")
     from gi.repository import Gdk, Gio, GLib, Gtk, Pango